+++ CHANGELOG.md
@@ -41,6 +41,7 @@
  - Set a default app_id (`managepy`) so you can use use gcloud compatible app.yaml files (which cannot contain an app_id).  Override with --app_id
  - Restricted access to the `clearsessions` view to tasks and admins only
  - Fixed the `sleep()` time in `djangae.utils.retry` which was sleeping in `ns` rather than `ms`
+ - Fix unicode error when creating a SQL representation
 
 ## v0.9.10